Let parser generate proper error messages
There is a sligth bug, as the incorrect line is shown when an error is
printed. I Haven't got time right now to check why this is the case, but
this eneds investigation. As an example: running the parser on a file
containing
```
var x = 1;
xor(
```
will generate the following error message:
```
bool.spl:2:4:
  |
2 | var x = 1;
  |    ^
unexpected (
```
The file position is actually correct, but for some reason the first
line is printed.