Use tab width of 4 in the lexer
This tab width is used by the lexer to calculate columns. While a tab is
only a single character, it will offset the column by more than 1. This
makes sure that the columns of lines using tabs, line up with other
lines. It's obviously a choice how much a tab should offset, but I think
most people agree it should be 4 (most editors do at least).

Megparsec however, thinks that 8 should be the default, resulting in
wierd error messages:
```
bool.spl:1:12:
  |
1 | 	xor(dfdfdfdf
  |            ^
unexpected (
```
You can see that it thinks the rror is in column 12, as there was a tab
in the source program.